requirements.txt
FastAPI
sqlalchemy
pydantic
uvicorn
pymysql
python-multipart
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
Pillow
